+++ src/modelling/PathPlanner.h
@@ -7,13 +7,13 @@
 
 class PathPlanner {
 private:
-  std::vector<Eigen::Vector3f> m_pathPoints;
+  std::vector<Eigen::Vector3d> m_pathPoints;
 
 public:
   PathPlanner();
-  std::vector<Eigen::Vector3f> getPath();
+  std::vector<Eigen::Vector3d> getPath();
 private:
-  void addPointToPath(const Eigen::Vector3f& nextPoint);
+  void addPointToPath(const Eigen::Vector3d& nextPoint);
   void setupPathPoints(int pointCount);
 
   void addCircleOfPointsToPath(int pointCount);
