+++ src/modelling/sensor-simulators/SensorSimulator3D.cpp
@@ -1,5 +1,5 @@
 #include "SensorSimulator3D.h"
 
-Vector3f SensorSimulator3D::getDataFor(int index) {
+Vector3d SensorSimulator3D::getDataFor(int index) {
   return m_data.at(index);
 }
